+++ plugins/services/sentry/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-02-09
+
+### Added
+
+- add self-hosted Sentry support to sentry plugin ([e3761ce3](../../commit/e3761ce3))
+
 ## [1.1.0] - 2026-01-30
 
 ## [1.1.0] - 2026-01-24